Sound overriding of methods with where constraints
Summary:
This is the followup to D3924359 to deal with sound overriding of
'where' constraints. As with constraints on generic parameters to methods,
the general principle is the following: any declared property of the overridding
method must be a consequence of declared properties of the overridden method,
because a caller might be calling through the superclass. The small complication
is that 'where' constraints that typically involve a type parameter might
be instantiated to become a general constraint between two types. For example:
```
class A { }
class Cov<+Tc> { }
class Base<Tb> {
  public function foo<Tf as A>(): void where Tb as Cov<Tf> { }
}
class Derived<Td> extends Base<Cov<Td>> {
  public function foo<Tf as A>(): void where Td as A { }
}
```
The `where` constraint in the superclass becomes
```
  Cov<Td> as Cov<Tf>
```
under instantiation (replace `Tb` by `Cov<Td>`). The type checker must show that
```
  Td <: A
```
is a consequence, which it can do, assuming `Tf <: A` as given by
the constraint on the generic parameter to the method.

Error messages are currently poor. See t13729267.

Reviewed By: oulgen

Differential Revision: D3974611

fbshipit-source-id: c7377ee5e78233747176bf22f395a9a1209776ff