Improve error reporting involving unnamed types
Summary:
Improve the synthetic names we use for types and members which do not naturally
have a name. Previous to this, an unnamed union would be given the name of its
first (named) member. This was done to match the ABI. However, this proved
confusing, as people think its referring to the actual member, not the
type. Instead, give it name like '(unnamed union containing 'foo')' to make it
clear that this name is synthesized. Also extend this logic to any unnamed type,
including structs. Furthermore, do not give unnamed members blank names, as this
is unhelpful. Instead give it a synthetic name indicating its an unnamed member
of a particular type.

Finally, standardize on using "unnamed" instead of "anonymous" as they're not
equivalent.

Differential Revision: D3989733

fbshipit-source-id: 932801a54c354ebb7bb6ef356e0d9fc5751e4b09