Define decl-based unification of types as mutual subtyping
Summary:
We're gradually moving towards a point where type checking is based only
on subtyping, and "unification" is replaced either by type equivalence
(i.e. types are subtypes of each other) or by the computation of an upper bound
on types (e.g. for merging of locals).

Decl-phase unification *is* just type equivalence, so that's how we define it
in this diff.

Reviewed By: dlreeves

Differential Revision: D3995060

fbshipit-source-id: edec70acca949da24387d829c6eae9e2796cfddc