Don't assume parent frame is the PHP caller in decodeCufIterHelper()
Summary:
This isn't true on PPC64 because they insert a dummy frame before every
call (to simulate their native calling convention).  This is the only
place we're making this assumption, and we can just as easily pass in
the frame pointer as an argument.

Also get the helper out of translator-runtime.

Reviewed By: swtaarrs

Differential Revision: D3979634

fbshipit-source-id: 052b5ce5de7b9001b832d3b5d3b17283cf44e468