Don't globally require MySQL
Summary: Now that it's possible to easily disable extensions, and the only place in the core of HHVM that relied upon MySQL is moved into the MySQL extension where it belongs, MySQL is no longer needed to compile HHVM itself. The async and normal MySQL extensions handle adding the includes and linking themselves.

Reviewed By: aorenste

Differential Revision: D2901049

Pulled By: Orvid

fbshipit-source-id: eff11ac83ffb453c0fc33ba78f9c38a2d23218f6