Get rid of Compiler::EnableHipHopSyntax
Summary:
I noticed that a number of tests which are supposed to run with
EnableHipHopSyntax are being compiled without it (they still run
with it turned on).

The reason is that the hphp_opts files were converted to .ini, and the
way the .inis were written, they just set the RuntimeOption, but when
compiling for repo mode we copy Option::EHHS to RuntimeOption::EHHS.

The ini parser doesn't really handle the compile time options properly
anyway, so the easiest fix was to just unify them.

While I was there I noticed that EnableXHP, and a slew of other options
were being copied back and forth, so fixed those too.

Reviewed By: swtaarrs

Differential Revision: D3985915

fbshipit-source-id: ce2eac6673a5d9352200677120679287f46731b6