Fix tc-print not mapping translations to bytecode
Commit c28f559 changed MD5's constructor to take a folly::StringPiece parameter
instead of a char*. This was causing OfflineTransData::loadTCData to call the
wrong MD5 constructor; hence an incorrect function md5 was being read from the
tcdata archive.
This diff solves the problem by adding a "size" parameter to MD5's general
"const void *" constructor,  making the conversion from char* to StringPiece
implicit. It also prevents the incorrect call of the general constructor when
passing any other type of pointer.
Fixes #7302.