Do not discared to_recheck1 and to_recheck2 sets in lazy incremental mode
Summary:
In lazy mode, I was trimming this set to only files that are open in IDE. The assumption was that during Full check we are always able to recover whatever was there by typechecking all of the dependencies.

But there is one case when this is not true: when we remove a definition from the file. When we do a Full_check, it's too late - the definition is not there, we don't know that it was there, so we won't report errors in a places where it is used. The initial redeclaration, when we went from "definitions is there" to "definition is not there", and the contents of to_recheck set then was the only moment when we were aware of this dependency.

To fix it, I keep the contents of this sets in environment during lazy check, and then merge them during a full check.

Reviewed By: jopamer

Differential Revision: D3985506

fbshipit-source-id: 7f94d6db0e1b8157e1aeeb18de82bb4fb7641226