Fine grained lazy incremental mode
Summary:
For initial simplicity, and due to a conflict that I solve in D3985481, lazy incremental mode was sometimes less smart than batch incremental mode. Incremental mode works fast by comparing old definitions to new ones and rechecking only the files dependent on things that have changed. But we were not doing it for definitions that were deferred to be redeclared "later" (by a full check) - instead we were just removing those definitions completely. This resulted in full check having nothing to compare against, and thus having to take ALL of the dependencies.

This diff changes it - we no longer remove those definitions, but move them to old heap, where they will wait for next full check. We now have to carefully keep track what is in what heap to avoid looking in wrong places.

Note: this trades performance inefficiency for slight memory inefficiency: if during the time when a definition is in the old heap some other command requires this definition, lazy decl will recompute it and put it in new heap. This lazy declaration is likely to happen in a place where we don't have access to all the global data that we need to update if we were to compare old definition to new one and discard the old one, so the definition will have to stay duplicated until next ServerTypeCheck.type_check. I hope that this is acceptable because:

- the things that we lazily require should be limited to dependencies of files open in IDE
- we already have this situation, in declaration phase of ServerTypeCheck.type_check, and it's our memory consumption peak. We're now just allowing this peak to possibly persist a little bit longer, instead of being resolved immediately.
- in the future diffs, I'll be working on heuristics that redeclare those things in idle times, so it will eventually be gone

Reviewed By: jopamer

Differential Revision: D3985527

fbshipit-source-id: 6a096e0f61fedb0262ac9c2e2da6ce3de79e1458