Error throttling for push diagnostics
Summary:
If you happen to make a change that introduces 10000 errors, you probably don't care about all of them anyway.

Right now we are using it only to avoid choking the editor with errors that it won't be able to display anyway, but on server side we always still compute full global error list when requested.

Note that in the future we can optimize it, by computing global error list in chunked batches, and stopping when we have more errors than we would sent to editor.

Reviewed By: oulgen

Differential Revision: D4004647

fbshipit-source-id: ad8ac26634aafadfd76d7c91787ce535df16db29