Fix a bug in incremental mode related to duplicate names
Summary:
We don't proceed to redeclaration for files that had naming errors. But by not doing it, we are missing the only moment where we can notice the situation described at the beginning of [[ diffusion/FBS/browse/master/fbcode/hphp/hack/src/server/serverTypeCheck.ml;8aa4623d98c9d1ed4679c8c9e9f4c1c8c4adcc8a$108-122 | add_old_decls ]].

I don't see a rationale for existence of this code - be removing it, at best, we'll do a no-op redeclare of few things. At worst, we'll actually switch the definition of duplicated class from one to the other and redeclare a bunch of things with this new definition. This is not incorrect though, as opposed to current behavior.

I think this bug has been there forever, all of those places blame to pre-2014.

Note: maybe this has something to do with Decl_heap_elems_bug? I couldn't think of repro case, but the result of this bug is that we are leaving some definitions hanging in shared memory with no pointers to it, which sounds bad.

Reviewed By: dlreeves

Differential Revision: D4026343

fbshipit-source-id: 9f1828b80ff4c3691414ce1d68c3201e5445067e