Run optimizedFCallBuiltin() for NativeImpl builtins
Summary:
The primary motivation here was to also optimize the inlined NativeImpl builtins
as we emit these in the JIT when we were able to statically determine that an
FCallBuiltin was safe but none was inserted by hphpc.

As an additional bonus some (but not all) of the optimizations were also safe
to perform in a non-inline context (basically all of the ones that didn't look
at the current context). These optimizations will now replace the CallBuiltin
which we emitted into their bytecode stubs.

Reviewed By: ottoni

Differential Revision: D3883189

fbshipit-source-id: d4577260428ec6926b633ee9e7d230049e33bac2