Use type_scan for req::root scanning.
Summary:
Each req::root is subclassed for a specific type. Use type_scan
infra for each type, instead of IMarker. Easiest way to plumb this
for now is keeping a virtual scan() method, but passing Scanner&
instead of IMarker&. The concrete root types can then look up
the type index, and we don't have to store it anywhere.

The alternative is recording the size & type index in the base
root_handle class, then scanning without making any virtual call.
But this moves some cost to when we construct the root, which is
undesireable.

Reviewed By: ricklavoie

Differential Revision: D3806539

fbshipit-source-id: 2eb026b47652b29e076bffb6d4a6fb371ff198df