Warn if watchman sync directory unspecified and use project root instead
Summary:
Using watchman needs a directory to which we can touch temp sync files.
This will warn the user if they haven't specified a sync directory in their hh config and falls
back to using the project root. It also attempts to touch a sync file during initialization
of watchman to ensure that we have the necessary permissions.

Reviewed By: jopamer

Differential Revision: D3991882

fbshipit-source-id: 124465cc6065f2ba60f130a2dbfc75d8f0250c0c