Factor isObj hacks into minstr-helpers.h (closer to jit helper apis)
Summary:
The various member instruction helpers in member-operations.h take
an isObj template parameter. When set, the TypedValue* base argument
is cast to ObjectData*. This hack is only used from JIT-helper
callers, as part of the table-dispatch pattern in minstr-helpers.h.

This diff factors the isObj template parameter out of member-
operations.h since it's not necessary except in the JIT. Once
this is done, it's relatively straightforward to split the
tables in minstr-helpers.h, to remove the isObj (and hacky cast)
completely. This diff only tackles the first part.

Reviewed By: swtaarrs

Differential Revision: D4032162

fbshipit-source-id: 6feca445ab6b5270041a716649d86a3e525ef1d6