Improve type checker coverage of bad void usage
Summary:
Void usage checking doesn't currently fire if the `void` type is inside a union type. This can occur with lambdas, as here:
```
function test_callback(): void {
  $f = $r ==> {
    if ($r === 3) {
      return;
    } else {
      return new C();
    }
  };
  $x = $f(3);
}
```
The fix is to iterate over Tunresolved types. Note that only the first bad type in a union is reported, to avoid duplication of errors.

Reviewed By: dlreeves

Differential Revision: D4044107

fbshipit-source-id: 812731e4f164ecc4d782cfd888604c93db097f84