Crank the main loop frequency from each 500ms to each 100ms
Summary:
In absence of client requests to serve, main loop does one iteration every 0.5 seconds, as controlled by select timeout in sleep_and_check. That means that it's the best granularity with which we can answer questions such as "are we idle?". I want to have finer granularity, in particular - start computing diagnostics 100ms since last activity.

This change means that we will do everything that we do periodically 5 times more often. It's not obvious if it's safe to check watchman (most likely not) or run ServerIdle (maybe...) that often, so I added separate timers to both of them to avoid changing their behaviour. The rest of the code of main loop //seem// fine.

Reviewed By: andrewjkennedy

Differential Revision: D4025198

fbshipit-source-id: b8c752b6da5ebc848c0d809fa749e5c385c288d4