Fix static invocations and member access on $$
Summary:
`$$` is not legal in the position that $x takes in `$x::someMethod()`; it breaks the parser. Once it parses correctly, we need to handle this as a special case in `make_class_id`.

Changing `make_class_id` doesn't affect the other callers, dynamic class instantiation (`new $$`), because we replace most names with `SN.Classes.cUnknown`, so the change here really only affects stuff in `Class_get` and `Class_const`, as we desire.

Fixes #7403.

Reviewed By: alexchow

Differential Revision: D4042646

fbshipit-source-id: d27350a8e4dfe6e241a64e16a5bcca1ade77a91f