Improve Vcost heuristic for inlining
Summary:
When deciding to inline a callee or not, one of the factors is the
"cost" of the inlined unit. This is done by generating the unit, and calculating
its "Vcost". Along with the rough number of instructions, this includes whether
or not the unit is "complete". Complete meaning that there's no bind or fallback
instructions in the main area. If the unit isn't complete, we assign it a
maximum weight, which effectively keeps it from being inlined. The idea is that
we don't want to inline callees which don't have a full region. However, this
completion heuristic considers Vasm instructions such as bindcc and fallbackcc
the same as bind and fallback. However, these instructions are typically used as
side-exits, and do not mean the unit is incomplete, even if on the main path. By
removing them we can inline more functions.

Reviewed By: paulbiss

Differential Revision: D3851561

fbshipit-source-id: e5b99947f4ba75da87932da666486cf78ceafdf9