Fix improper parser constant folding and hhbbc crashes on Hack arrays
Summary:
If you compare two Hack array literals, the parser will attempt to
constant fold the comparison during parsing (just as it does with normal
arrays). This causes two problems. First, since getScalarValue() on an
ExpressionList always returns an array, we'll inadvertantly perform key
conversion on the Hack arrays before the comparison (even though we convert the
array to the proper Hack array type). Thus, the constant folded comparison may
be incorrect. The second problem is that certain Hack array comparisons can
throw a PHP exception. This is supposed to work, but we might attempt to JIT the
PHP code, which has lock ordering issues if we're in sandbox mode.

Fix these issues by adding a new iterface to ExpressionList to obtain the scalar
values in the list via a callback mechanism. This lets us build the proper array
type directly, rather than always going through a PHP array, avoiding key
conversion. Fix the second issue by making sure that the JIT is disabled before
starting file parsing.

Finally, hhbbc has a similar issue where it can attempt to constant reduce Hack
array comparisons in the bytecode and end up attempting to run PHP. Add a helper
function which wraps the constant reduction in a try-catch and inhibit the
reduction if an exception is thrown.

Reviewed By: markw65

Differential Revision: D4049866

fbshipit-source-id: 024d73704eabca6f5723fd75be4984938c05bf6f