Support applying local changes to SharedMem
Summary:
There are cases when we want to apply speculative changes over the view of the shared heap. Currently the way to do this is to utilize some combination of oldify/revive or shelve/unshelve and remove. This is error prone because if we fail to leave the world in the previous state then this can lead to invariants/assertions being violated, causing the server to crash.

This diff introduces the notion of a stack of local changes that can be applied on top of the heap. Modifications like add/remove/move are performed on a hash table kept on the local process, instead of on the shared heap. Read operations like mem/get will check for modications in the local stack, if none affect the key then the next stack is queried ultimately ending with reading from the shared heap. Since the changes are kept local, it is easy to go to the previous view of the world.

Depends on D4024202

Reviewed By: dabek

Differential Revision: D4026151

fbshipit-source-id: 5de2c56b80f1d598c7d57b7d6ff2be954fb7a088