Use local change stacks in ServerIdeUtils
Summary:
ServerIdeUtils utilizes shelve/unshelve to apply temporary changes to shared memory before reverting it. This logic is complicated because we don't do a full redeclaration so we need to be careful to revive class elements that may be removed.

Using LocalChanges we can avoid this complexity. We push a new stack of changes, perform type checking and then pop off the stack. By definition this will leave the world in the identical state. This should get rid of the decl elements crash that is sometimes triggered during IDE commands.

Reviewed By: dabek

Differential Revision: D4056178

fbshipit-source-id: 79f39edb5d0c852919becfe8acdb65d14a9fa295