Update expected errors in test case
Summary:
D4024506 and D4026343 depend on each other, but were not actually stacked on my devserver when I created them, so the test output in one of them got out of date.

The only difference is that we re-analyze bar_expects_int.php once again (and discover that the errors are the same, but we push them once again anyway). The re-analyzing of this file is a direct consequence of D4026343.

Reviewed By: jamesjwu, oulgen

Differential Revision: D4076304

fbshipit-source-id: 0d0f444772d25959dce94737bd744bad66020a44