Save insts to delete while we are iterating over preds in DCE.
Summary:
removeUnreachable started crashing in prod recently.  I haven't been able to get any repro steps but I've looked at several core files.  According to the cores, "it" is pointing at the end of the preds list at line 206 where the code is trying to increment the iterator.  I'm not sure how "it" is pointing to end() at this point since the code seems to guard against this.  In all the cases I've looked at so far, the lists always seem to have only two elements left.

I know it should be safe to delete items from a list as long as the iterator is not pointing at the erased element but I've rewritten the code to save all the insts that need to be deleted to process them after we are finished iterating.  This will avoid any possible iterator invalidation issues.  If that doesn't fix the crashes we can revert this change.

Reviewed By: markw65

Differential Revision: D4077161

fbshipit-source-id: df72326188a798ec3986098a5940a5c1489d5b5a