Basic framework to JIT all optimized translations at once using hfsort's function order
Summary:
This diff adds a basic framework to trigger JITing all optimized translations at
once and to place the functions in the TC according to hfsort's call-graph-based
algorithm.  The triggering mechanism is simply based on the number of requests
executed, which is specified by Eval.JitRetranslateAllRequest.  Setting this
runtime option to zero disables the mechanism entirely, and this is still the
default.

The 3 basic steps in this mechanism are:
1) Generating optimized translations for all profiled functions;
2) Sorting the functions using hfsort;
3) Relocating the functions in the TC according to the selected order.

This diff implements step 2 only.  This step involves both creating a weighted
call graph for all the functions that were profiled and running hfsort on it.
In order to obtain a weighted call graph, this diff extends ProfData to keep
track of the translation containing each PHP call (for profiling translations),
and uses this information in combination with the prologue's callers information
that was already saved.

Reviewed By: swtaarrs

Differential Revision: D4026017

fbshipit-source-id: 9574f823abe24b2677ed2c2144b222473624f584