Start generating a javascript API for the full-fidelity tree
Summary:
Generating classes for a full-fidelity JavaScript API.

So far this is just classes for an "editable" tree; that is, a library of Javascript classes that supports rewriting.

The data structures here are intended to be used in an immutable, persistent manner.

I've given a sample rewriter, that turns all comments into "blah blah blah".  I'll come up with more sample rewriters in a later diff.

Sorry this is so huge, but it is mostly generation of boilerplate.

Reviewed By: dlreeves

Differential Revision: D3976787

fbshipit-source-id: b49ba69b27329f597c859fe4b88a45af15928f31