Fix some parser naming issues
Summary:
The parser attempted to track the name of the current function, and
the name of the function containing the current closure, but it didn't
take account of nested functions. This meant that m_funcName and
m_containingFuncName could be stale at the point of use.

In most cases, this just resulted in the incorrect names being
displayed in backtraces, but in a couple of cases, we could sneak a
parse time fatal past the parser, by making it think we weren't in a
special method.

Reviewed By: swtaarrs

Differential Revision: D4075313

fbshipit-source-id: bd891b35590953e0c17c86c0af50f97cb0996df3