Fix closure and anonymous class naming issues
Summary:
We relied on them having unique names, but didn't guarantee that they
were unique.

We also defined them when the unit was loaded, which meant that
instantiating a unit that defined many closures would allocate memory
for the classes, even if they were never used. Also, its possible that
an anonymous class depends on a class or interface that isn't defined
yet (eg if its defined in a function).

Reviewed By: swtaarrs

Differential Revision: D4078637

fbshipit-source-id: c0143d80c22614ffd6c41bc6f0a233ab4df2860d