Fix lexing for doc string literals
Summary:
The spec stated that a doc string literal body must be at least a single newline, but the spec is wrong. It is legal for a doc string literal to simply be

<<<FOO
FOO

With no body line between the header and the trailer.

I've submitted a pull request to update the spec with the correct grammar.

Reviewed By: dlreeves

Differential Revision: D4046077

fbshipit-source-id: 47a96a97c7419308bd5b5739e5e6323e3720b5a4