Add attribute to enum
Summary:
An enum declaration may begin with an attribute specification. This detail was not in the specification. I've submitted an update to the spec.

If you look at the diff to the test case you'll see that the attribute spec body is a list item, which is inconsistent with the other parse tree nodes that have singleton lists.  The attribute list parser was never updated when I added new list processing code. I'll fix that up in a later diff.

Reviewed By: savil

Differential Revision: D4051480

fbshipit-source-id: 7f5805ad6c9c7dc7a0786521ac8be28cb64557f3