Fix some closure issues
Summary:
Certain Reflection operations need the closure class to be marked loaded in rds.
I was trying to do that in the closureInstanceCtor, but that only worked if the first
closure that gets instantiated is the template class.

Fix it to work no matter which closure class is instantiated, and
optimize for repo mode, where we know the closure will be persistent.

Reviewed By: swtaarrs

Differential Revision: D4095461

fbshipit-source-id: 3c1ac3c45f05f3ab068d885d6847dfa7fbea26dc