Add a heuristic to disambiguate cast operators from paren exprs.
Summary:
The original Hack parser has an inconsistent heuristic for disambiguating cast operators from parenthesized expressions; for example it treats

(x) | $y

as a cast but

(x) ^ $y

as a parenthesized expression.

In this diff I propose and implement a more well-founded heuristic. I will also update the github specification.

There is still work to be done to determine (1) whether casts should be allowed to be more than simple names, and (2) how casts interact with XHP.

Reviewed By: dlreeves

Differential Revision: D4068809

fbshipit-source-id: 657757c0fbb4c7e4258cb14de676031096f01350