[Aarch64] Fix decqmlock with acquire and release semantics
The VASM decqmlock and incqmlock have no ordering semantics when
using ldxr and stxr Aarch64 instructions. The current code is
broken and could be fixed by adding DMB barrier instructions,
however the cleaner approach taken here is to use the load-acquire
and store-release, which is equal to x86 including ordering
characteristics. This is also the same as PPC64's decqmlock.