Suppress duplicate and cascading parse errors
Summary:
The full-fidelity parser sometimes reports exact duplicates of an error; we now filter those out of the error list in the syntax tree.
We also wish to suppress "cascading" errors, such as reporting both an "expected expression" and "expected semicolon" error at the exact same location if you have something like "x = }".
I've added an option to the syntax tree and parser service to remove duplicated errors.

This is just the first of several changes I plan to make to improve error reporting; I'll check in a test case in a later diff.

Reviewed By: jamesjwu

Differential Revision: D4082745

fbshipit-source-id: b1f0979d7dc7631dfcdbd122b91fda1090d61d3f