Actually support all the curl constants that were added
Summary:
Way back when in D3056455, I added the code to register a few hundred Curl constants that we hadn't registered. Unfortunately, I forgot to tell `CurlResource` about them, so you couldn't actually use any of them.
This informs `CurlResource` about all of the constants I added.

This also adds support for setting options that are allowed to be set to null, to null. Previously we were treating them as normal string options.

Reviewed By: aorenste

Differential Revision: D4108246

fbshipit-source-id: c7ccaa3e01b813887c1f87ced47513e471e56190