Properly fatal when redefining builtin functions
Summary:
Attempting to redefine a builtin function should fatal. However, in sandbox mode
doing so triggers an assert in debug builds. In prod builds, it silently accepts
the redefinition! In repo mode the redefinition is silently dropped. None of
these behaviors are desirable, so ensure we fatal instead.

In repo mode, when the parser sees a function definition with the same name as a
builtin, emit a Fatal bytecode in place of the function definition. If the
definition is at the top-level, the unit fatals immediately before running
anything (this matches Zend behavior). If the definition is conditional, all the
code up to the definition is executed.

Fataling in sandbox mode is just a matter of checking if the function is
already defined instead of assuming it is, if its persistent.

Reviewed By: markw65

Differential Revision: D4099377

fbshipit-source-id: c806fb6f337a6e607d3450c911f85dbfc63f624c