Add a JitMaxProfileTranslations runtime option
Summary:
And use it to limit the length of retranslation chains for Profile translations.
The JitMaxTranslations option is still used to limit the length of retranslation
chains of Live translations. This allows us to independently limit the length of
the retranslations chains for these two kinds of translations.  Also tune both
options.

Reviewed By: swtaarrs

Differential Revision: D4121520

fbshipit-source-id: cb91a4988946ad7f93d7175f849fe169e8a90938