Delete Check* -> Jmp shortcut in preOptimizeCheckLocation()
Summary:
We expect that the type of the location or value is appropriately
refined after we gen any Check* instruction, but this optimization makes
it not so.  This should already be covered by load-elim, so just stop
doing it here.

This causes us to produce more TBottoms than before, which we handle
gracefully except for some asserts, which are also amended in this diff.

Reviewed By: swtaarrs

Differential Revision: D4123864

fbshipit-source-id: 25593f73c1b86202ac65fbc8f3bbc4f4d21d4b08