Update BCMarker before generating inlined function return block
Summary:
When generating the return block for an inlined function, update the
current BCMarker to point to one of the Ret* bytecodes in that
function. Otherwise, the marker for all the IR instructions generated in that
block will point to whatever bytecode we happened to last process. In most
cases, that will happen to be one of the Ret* bytecodes in the function, but if
the function contains loops it may not be. This can cause problems because we
may not then have accurate stack depth information. Fix this by explicitly
setting the marker to the same as the last IR instruction in one of the return
block's predecessors.

Reviewed By: mxw

Differential Revision: D4130095

fbshipit-source-id: 7d029be16aee740e41f01f9e48029e39bf7fcc5f