Fix expected output for private_props_020.php
Summary:
When we inline and elide the frame, we lose the inlined frame in
backtraces.

When we disable pgo, we inline the constructor; with pgo enabled, we
only call the constructor once, so we don't get the chance to inline;
but running with --count=2 (in dbgo builds) the error is reported on
line 18 on the second run.

Reviewed By: ricklavoie

Differential Revision: D4134805

fbshipit-source-id: d5ba2af26f5bdf4e9ad04b13251789b3694791eb