Prepping the front-end of `hh_single_parse` to support multiple parsers
Summary:
We want to use `hh_single_parse` the way we use it now, but we also want to use
it as a front-end to the full fidelity parser.

This diff extends the front-end by adding a choice of parser (using `--parser`)
but also allows you to compare results of both the old and the new parser. I
found this useful for debugging (see test plan).

Reviewed By: jamesjwu

Differential Revision: D4104455

fbshipit-source-id: c257eaba9965b61b81e9099f2bc018aee271e119