Conservative-scan partially initialized apc objects
Summary:
When doing apc_fetch of an object, we have a fast path
that initializes properties without first pre-initalizing
them all to default values. It's possible to trigger GC
while doing this (classes with static properties re-enter).
Even though the half-initialized objects aren't visible to
PHP, they're reachable and need to be scannable.

Reviewed By: kodafb

Differential Revision: D4124459

fbshipit-source-id: f60ff477c1ebe1d24fc72d5f010077a940c2d6a8