Oldify deferred defs before phase2 redeclaration
Summary: Switching the order from "redecl and oldify" to "oldify and redecl" - this is needed because the redecl might depend on some of the definitions we want to oldify, and if we don;t do it first, it will use outdated version. Since the redecl can now process some of the oldified things, we also need to update how we track what is oldified (needs_phase2_redecl).

Reviewed By: jopamer

Differential Revision: D4129052

fbshipit-source-id: a8818c10ef7636cdfc1e6b0178637cdca26589bf