Limit "stop at parsing errors" behavior to IDE files only
Summary: As predicted in {D4026151}, the fact that LocalChanges doesn't work in parallel code is error prone. Here, the set of "disk_files" can be processed by parallel workers, in which case everything breaks. Changing it to only depend on LocalChanges for handling ide_files (which are always processed by the master... for now).

Reviewed By: jopamer

Differential Revision: D4129059

fbshipit-source-id: 52c00092f831de3a7a601d2caadc052626acf924