Include phase2redecl in to_recheck set
Summary:
There is an explanation why we can skip it in lazy decl mode, but I feel that it's
very brittle and hard to reason about. There is no cost in doing it on both
paths (lazy check will just add this set to env.needs_recheck), and it simplifies the code a bit.

Reviewed By: jopamer

Differential Revision: D4129065

fbshipit-source-id: efb0afec7faa60e9aad35c9a8f681f8446e13600