Retarget guard jccs to a closer, compatible jcc when possible
Summary:
See comments in the code for details. In an example unit with 3 Int
guards and one Uncounted guard, this cuts the guard sequence down from 42 bytes
to 28 bytes:

  0xaa0027a:  cmpb  $0x11, -0x8(%rbp)
  0xaa0027e:  data16 nop
  0xaa00280:  jnz 0xe600835
  0xaa00286:  cmpb  $0x11, -0x18(%rbp)
  0xaa0028a:  jnz 0xe600835
  0xaa00290:  cmpb  $0x39, -0x28(%rbp)
  0xaa00294:  jnle 0xe600835
  0xaa0029a:  cmpb  $0x11, -0x38(%rbp)
  0xaa0029e:  jnz 0xe600835
  0xaa002a4:  ...

becomes:

  0xaa0027a:  cmpb  $0x11, -0x8(%rbp)
  0xaa0027e:  jnz 0xaa00290
  0xaa00280:  cmpb  $0x11, -0x18(%rbp)
  0xaa00284:  jnz 0xaa00290
  0xaa00286:  cmpb  $0x39, -0x28(%rbp)
  0xaa0028a:  jnle 0xaa00290
  0xaa0028c:  cmpb  $0x11, -0x38(%rbp)
  0xaa00290:  jnz 0xe600835
  0xaa00296:  ...

The presence or absence of that `data16 nop` is dependent on the alignment of
the whole unit, but most of the savings come from the 4 bytes shaved off of
each of the 3 retargeted jccs.

Reviewed By: ottoni

Differential Revision: D4136117

fbshipit-source-id: 6555d20fd29c43aeea1477b6a0facb185943bb6a