Fix type hole with new on union of classnames
Summary:
Hack does not properly check the construct
```
new $t(arg1,...,argn)
```
if `$t` is a union of `classname<t>` types. For example, in the code
```
<<__ConsistentConstruct>>
class C {
  public function __construct(int $s) {}
  }
final class D {
  public function __construct(string $f) {}
}

function test_it(bool $b): void {
  if ($b) {
    $t = C::class;
  } else {
    $t = D::class;
  }
  $x = new $t(2);
}
```
the type-checker does not notice that `$t` might be `D::class`
in which an integer is passed to the constructor that expects
a string.

The fix is to check constructor calls across the union type, and
produce a union of results.

Reviewed By: dlreeves

Differential Revision: D4124394

fbshipit-source-id: 1f14cbdfe8304019e5a47eea17ed3960141fefc3