Guard on the value of the member base register
Summary:
Right now, as an optimization, we avoid breaking tracelets in the middle
of a minstr stream because we only track information about the member
base at irgen-time.

We can do better by adding Check instructions for the member base value
and doing proper tracking (analogous to that for locals and stack slots)
in FrameStateMgr.  This also lets us generate better PGO'd regions when
before we would not have had any type information for the member base.

Reviewed By: swtaarrs

Differential Revision: D3096032

fbshipit-source-id: e215a6bbe381fef9e96382c39e16c9f2bd798e45