Selectively relax DataTypeCountness to DataTypeGeneric
Summary:
Add back DataTypeCountness as the datatype category for values that are simply
IncRef'd/DecRef'd, regardless of being boxed.  Use this in a few places,
primarily for the local DecRefs at the end of the functions and when pushing and
popping values on the VM stack.  In addition, this diff also extends the
PGO-based guard relaxation to selectively relax DataTypeCountness to
DataTypeGeneric.

Reviewed By: paulbiss

Differential Revision: D4137092

fbshipit-source-id: 055dc9b4fb68c03c9776e60d52586e6f731078e2