Remove TypedValue unions from ActRec
Summary:
dummyA and dummyB are never used to actually hold TypedValues.
m_r was used, but kinda not really - the address of m_r is where
the returnvalue is placed when we tear down an ActRec, but what
the VM is logically doing is pop(ActRec)+push(TypedValue), so it's
not critical that we declare m_r as part of ActRec.

Reviewed By: swtaarrs

Differential Revision: D4169749

fbshipit-source-id: 556049f124df90e4175c68008fc44da3b97d121c