Introduce the disk heap
Summary: Introduces a new sharedmem structure to keep track of disk and IDE changes. This allows the ParserHeap to always grab the latest version of a file once it needs to start lazily reading from disk in a later diff. For now, this doesn't increase our memory usage, since we only use it on quick parsing.

Reviewed By: dabek

Differential Revision: D4084519

fbshipit-source-id: ff1a920c20629bb89f0a7cbb7a26199cbe9535c8