Create a Hack API front end to the full-fidelity parser
Summary:
Create a hack API front end to the full-fidelity parser

Right now I want this to be reviewed to see whether this API makes sense for the real-world scenarios.  We'll get proper tests, etc, going in a later diff.

This is not debugged yet and there are likely many small errors still.

Differential Revision: D4134093

fbshipit-source-id: e752fd44a88e7fab8e3d8d9f80c6c438e5f0da84